backend/requirements.txt
bcrypt
beanie
fastapi
httpx
pydantic-settings
PyJWT
python-multipart
uvicorn
